requirements.txt
channels==3.0.5
Django==3.2
mysqlclient==1.4.4
pypinyin==0.46.0
Pillow==9.2.0